refactor: move `Filter` to `messages`
